refactor: replace swing components to javafx
Refactor the whole project through replacing the swing components with
javafx if possible, which includes:
 - java.awt.Color            to  javafx.scene.paint.Color
 - java.awt.Graphics2D       to  javafx.scene.canvas.GraphicsContext
 - java.awt.geom.Point2D     to  javafx.geometry.Point2D
 - java.awt.geom.Rectangle2D to  javafx.geometry.Rectangle2D
 - javax.swing.Timer         to  javafx.animation.AnimationTimer

It is the first step of refactor and does not include any design pattern,
it only focus on replace swing with javafx. However, there still remain
some parts of code(keyboard reaction) that I'm not capable to
handle/refactor, which may needs my further study into javafx.